requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
anthropic>=0.7.0
mss>=9.0.1
pillow>=10.0.0
psutil>=5.9.0
pydantic>=2.0.0
pathlib2>=2.3.7